update ssl and nll
rename net_run_noise as net_run_nll
update ssl_abstract
